Removed un-needed changes, reorganized file structure and added comments
