+++ lib/cloudinary.js
@@ -3,7 +3,7 @@ exports.config = require("./config");
 exports.utils = require("./utils");
 exports.uploader = require("./uploader");
 exports.api = require("./api");
-let account = require("./account");
+let account = require("./provisioning/account");
 
 exports.provisioning = {
   account: account,
