+++ lib/config.js
@@ -44,7 +44,7 @@ function putNestedValue(params, key, value) {
   return params;
 }
 
-function getParsedConfigFromEnv(ENV_STR) {
+function parseCloudinaryConfigFromEnvURL(ENV_STR) {
   let conf = {};
 
   let uri = url.parse(ENV_STR, true);
@@ -70,21 +70,21 @@ function getParsedConfigFromEnv(ENV_STR) {
   return conf;
 }
 
-function enrichConfigWithURIQuery(ENV_URL) {
+function extendCloudinaryConfigFromQuery(ENV_URL, confToExtend = {}) {
   let uri = url.parse(ENV_URL, true);
   if (uri.query != null) {
-    entries(uri.query).forEach(([key, value]) => putNestedValue(cloudinary_config, key, value));
+    entries(uri.query).forEach(([key, value]) => putNestedValue(confToExtend, key, value));
   }
 }
 
-function enrichCloudinaryConfig(parsedConfig) {
+function extendCloudinaryConfig(parsedConfig, confToExtend = {}) {
   entries(parsedConfig).forEach(([key, value]) => {
     if (value !== undefined) {
-      cloudinary_config[key] = value;
+      confToExtend[key] = value;
     }
   });
 
-  return cloudinary_config;
+  return confToExtend;
 }
 
 module.exports = function (new_config, new_value) {
@@ -99,9 +99,11 @@ module.exports = function (new_config, new_value) {
     let CLOUDINARY_ENV_ACCOUNT_URL = process.env.CLOUDINARY_ACCOUNT_URL;
 
     [CLOUDINARY_ENV_URL, CLOUDINARY_ENV_ACCOUNT_URL].forEach((ENV_URL) => {
-      let parsedConfig = getParsedConfigFromEnv(ENV_URL);
-      enrichCloudinaryConfig(parsedConfig);
-      enrichConfigWithURIQuery(ENV_URL);
+      let parsedConfig = parseCloudinaryConfigFromEnvURL(ENV_URL);
+      extendCloudinaryConfig(parsedConfig, cloudinary_config);
+      // Provide Query support in ENV url cloudinary://key:secret@test123?foo[bar]=value
+      // expect(cloudinary_config.foo.bar).to.eql('value')
+      extendCloudinaryConfigFromQuery(ENV_URL, cloudinary_config);
     });
   }
   if (!isUndefined(new_value)) {
