+++ lib/provisioning/account.js
@@ -1,5 +1,5 @@
-const config = require("./config");
-const call_account_api = require('./network/call_account_api');
+const config = require("../config");
+const call_account_api = require('../network/call_account_api');
 
 /**
  * @desc - List all subAccounts under the account.
